Now k=6, seasons out, null model for comparison, updated regions
1. Combined NEUS into one "region"  [note I did not update the
projection code so may find an issue there]
2. Increased the "k" for temperature terms to 6. This solves the wacky
cod fit where cod were predicted to increase at very high temps. This
only happened in the multi-region models (not in any of the NEUS-only
models I fit for our other project), so must have been an artifact of
modeling observations from disparate regions/seasons/thermal
conditions.
3. Added code to create figures of the gam fits for easy checking - to
make sure the smooths do not go crazy like the cod model did. Note that
I did not implement an automated checking like we discussed, as the
fits are just too different by species to come up with a general rule
that works.
4. Added code to fit gams without temperature terms in order to compare
the deviance explained by models with and without temperature. I think
we should use this as a criterion for selecting which species to model.
If adding temperature terms only explains <5% of the deviance, I'm not
sure it is meaningful to do the projections. 5% is an arbitrary cutoff.
These terms are now output in the diagnostics table.