Allow species projections outside historical regions
@larogers, this may be of interest (and I’d be curious what you think).

I modified the species range projection code so that species are not
constrained to the regions used for fitting the SDM. A standard region
intercept and mean biomass is used for projection into all regions
(only temperature and rugosity vary).

Added a new flag called projtype, which can be set to ‘_xreg’ for
cross-regional projections.

The biomass_proj_mapscontinent_*.pdf map shows all regions together (by
ocean). The observations are plotted on the left as well. While not
perfect, the projections don’t look half bad.