Updates to all model fits across all regions in an ocean
This assumes that all species would be recorded if they were found in
all regions. The benefit is that it allows a better estimate of thermal
preferences (using full range from the continent, not just the range
within the regions in which it was observed).

Required some gymnastics in 6_model_fitting_loop to deal with regions
with zero presences. Script now replaces the region label in regions
with no presences with the region label from the region with the most
presences. This prevents the model from fitting a very low region
coefficient to explain the lack of presences. The script also adds a
few fake non-zeros (logabundance= -23) to these regions so that they
can be included in the abundance model fit.

The models now only have a linear biomassmean term, rather than
smoothed.

PDF of model terms now includes all terms

In 7_rangeprojection.R, small updates to deal with the fact that some
regions included in the model projections will have zero presences.