Making grid connection points for NatCap
Grid connection points and landing points are needed for NatCap Wave
model (and can be used by the wind model). The code uses towns >1000
people as proxies for grid connection points, and picks the closest
point on the coast as the landing point.

Also renamed make_NatCap_AOI.r to make_NatCap_files.r to reflect the
broader purpose of the file